requirements.txt
praw
pandas
yfinance
python-dotenv